requirements.txt
python-dotenv
streamlit
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
PyMuPDF
docx2txt
-e .